Разбиение проекта на пакеты по фичам + исправлено направление в графе навигации на от search к vacancy
