+++ app/src/main/java/ru/practicum/android/diploma/filters/ui/views/CountryFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.placeofwork
+package ru.practicum.android.diploma.filters.ui.views
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import ru.practicum.android.diploma.R
 
-class PlaceOfWorkFragment : Fragment() {
+class CountryFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return inflater.inflate(R.layout.fragment_place_of_work, container, false)
+        return inflater.inflate(R.layout.fragment_country, container, false)
     }
 }
