+++ app/src/main/java/ru/practicum/android/diploma/filters/ui/views/IndustryFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.country
+package ru.practicum.android.diploma.filters.ui.views
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import ru.practicum.android.diploma.R
 
-class CountryFragment : Fragment() {
+class IndustryFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return inflater.inflate(R.layout.fragment_country, container, false)
+        return inflater.inflate(R.layout.fragment_region, container, false)
     }
 }
