+++ app/src/main/java/ru/practicum/android/diploma/filters/ui/views/PlaceOfWorkFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.team
+package ru.practicum.android.diploma.filters.ui.views
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import ru.practicum.android.diploma.R
 
-class TeamFragment : Fragment() {
+class PlaceOfWorkFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return inflater.inflate(R.layout.fragment_team, container, false)
+        return inflater.inflate(R.layout.fragment_place_of_work, container, false)
     }
 }
