+++ app/src/main/java/ru/practicum/android/diploma/search/ui/SearchFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.search
+package ru.practicum.android.diploma.search.ui
 
 import android.os.Bundle
 import android.view.View
