+++ app/src/main/java/ru/practicum/android/diploma/vacancy/ui/VacancyFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.vacancy
+package ru.practicum.android.diploma.vacancy.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
